+++ myCookingManager/frontend/src/components/RecipeDetailsCategorySelect.jsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from "react";
 import { styled } from "styled-components";
 import { useAuth0 } from "@auth0/auth0-react";
 
-import { RecipeDetailsContext } from "./RecipeDetailsContext";
+import { RecipeDetailsContext } from "./contexts/RecipeDetailsContext";
 
 import CategoryCreation from "./CategoryCreation";
