+++ myCookingManager/frontend/src/components/RecipePreview.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useRef } from "react";
 import { styled } from "styled-components";
 
-import { CatalogueFlowContext } from "./CatalogueFlowContext";
+import { CatalogueFlowContext } from "./contexts/CatalogueFlowContext";
 import useAutoScrollIntoView from "../utility_functions/hooks/useAutoScrollIntoView";
 
 import DialogueBox from "./DialogueBox";
