+++ myCookingManager/frontend/src/components/contexts/Auth0ProviderWithNavigate.jsx
